fix: resolve code review issues across geocoding, bookmarks, music player, and calendar
- geocoding: remove cfg gating from geohash_to_coords for cross-platform
- geocoding: only cache successful responses, not network errors
- geocoding: use safe coordinate access to prevent panics on malformed data
- bookmarks: add generation token to prevent stale NIP-51 publishes
- bookmarks: replace tokio::time::sleep with platform timer abstraction
- music_player: release lock before storage::set to avoid blocking
- event_map: remove unused clone and move generation bump earlier
- use_unsaved_changes: remove unconditional unregister in use_drop
- mobile.rs: add SAFETY comment for unsafe JObject construction
- build-android.sh: add ShellCheck SC2064 disable for trap expansion
- MainActivity.kt: use AssetFileDescriptor for reliable file size check
- clippy: fix duplicated cfg attributes and use first() instead of get(0)

+++ src/components/calendar/view.rs
@@ -532,7 +532,7 @@ fn position_day_events(events: &[UnifiedEvent], _date: &str) -> Vec<PositionedEv
             {
                 use chrono::{Local, TimeZone, Timelike};
                 // Clamp timestamp to max safe value (9999-12-31) before i64 cast to prevent overflow
-                let ts_clamped = (ts as u64).min(253_402_300_799) as i64;
+                let ts_clamped = ts.min(253_402_300_799) as i64;
                 let dt = match Local.timestamp_opt(ts_clamped, 0).single() {
                     Some(dt) => dt,
                     None => {
@@ -789,7 +789,7 @@ fn format_event_time(event: &UnifiedEvent) -> String {
     {
         use chrono::{Local, TimeZone, Timelike};
         // Clamp timestamp to max safe value (9999-12-31) before i64 cast to prevent overflow
-        let ts_clamped = (ts as u64).min(253_402_300_799) as i64;
+        let ts_clamped = ts.min(253_402_300_799) as i64;
         let dt = match Local.timestamp_opt(ts_clamped, 0).single() {
             Some(dt) => dt,
             None => {
